Large Credit Cooperatives of Paraguay

The py_creditcoops data frame has 44 rows and 3 recorded columns. The observations are total loans granted from large Paraguayan credit cooperatives (Type A) paired for the years 2016 and 2018.

This data set was used in a published paper analyzing Paraguay's large credit cooperatives.

Usage

py_creditcoops

Format

A data frame with 44 rows and 6 variables:

rowid

index

country

Country name

ISO

Country ISO code

coop_id

Unique identifier of the cooperative

year

A factor indicating the year the loan was granted

total_loans

A numeric variable indicating total loans granted in US dollars

Source

National Cooperative Institute (INCOOP). https://www.incoop.gov.py/

Note

The respective credit unions have deliberately not been named, however, can be requested from the author.

References

Schneider, A. (2022).Kreditgenossenschaften: Marktstruktur, Wettbewerb und Verhalten. Das Beispiel Paraguay, Zeitschrift für das gesamte Genossenschaftswesen, 72, 7-38, https://doi.org/10.1515/zfgg-2022-0002.