All-cause mortality data

The py_deathcauses data frame has 2520 rows and 3 recorded columns. The observations are counts for 21 disease categories on a national level for the years 2013 and 2022.

This data set was used in a published paper analyzing all-cause mortality during the first total lockdown in the COVID-19 pandemic era.

Usage

py_deathcauses

Format

A data frame with 2520 rows and 6 variables:

rowid

index

country

Country name

ISO

Country ISO code

date

Date variable in %Y-%m-%d format

category

A character string with a disease

counts

A numeric variable indicating the total counts of a given disease in a given time period

Source

Ministry of Public Health and Welfare (MSPBS). https://www.mspbs.gov.py/index.php

References

Schneider, A. (2024).The impact of the COVID-19 pandemic on all-cause mortality in Paraguay: 2020 - 2022, Revista científica ciencias de la salud, 6, 1-7, https://doi.org/10.53732/rccsalud/2024.e6155.