Monthly street crime

The py_streetcrime data frame has 1248 rows and 11 columns. The observations are monthly counts for five typical street crimes such as theft, aggravated theft, robbery, aggravated robbery, and aggravated robbery resulting in death time series. The data correspond to Asuncion (central branch) and their respective branches (barrial1, barrial2 etc.), as well as the cities of Concepcion, Encarnacion, and Ciudad del Este. The definitions of these crimes are according to the Paraguayan Penal Code. The time series cover the years 2014 to 2021.

This data set was used in a published paper analyzing street crime during the first total lockdown in the COVID-19 pandemic era in Paraguay.

Usage

py_streetcrime

Format

A data frame with 1248 rows and 11 variables:

rowid

Index

country

Country name

ISO

Country ISO code

branch

Branch of the Public Ministry

area

Administrative area of the branch

year_month

Date variable in %Y-%m-%d format

theft

Count value corresponding to the respective time period and branch.

theftagg

Count value corresponding to the respective time period and branch.

robbery

Count value corresponding to the respective time period and branch.

robberyagg

Count value corresponding to the respective time period and branch.

robbery_death

Count value corresponding to the respective time period and branch.

Source

Public Ministry (MP). https://www.ministeriopublico.gov.py/

References

Schneider, A. (2024). Street crime during COVID-19: a retrospective comparison of two Latin American capitals, Revista Paraguay desde las Ciencias Sociales, 14, 73-91, https://publicaciones.sociales.uba.ar/index.php/revistaparaguay/article/view/9800.