Weekly mean temperatures and holidays

The py_temperature data frame has 1565 rows and 7 columns. The observations are weekly average temperature data for the five largest Paraguayan cities from 2016 to 2021. The data set also includes an indicator variable if the week had a holiday.

Usage

py_temperature

Format

A data frame with 1565 rows and 7 variables:

rowid

index

country

Country name

ISO

Country ISO code

city

Factor, city name

week

Week, Y%-m%-d% format

avg_temp

Average weekly temperature at the location

holiday

Dummy variable if the respective week had a holiday

Source

Visual Crossing. https://www.visualcrossing.com/