Daily weather data

The py_weather data frame has 10945 rows and 28 columns. The observations are daily weather conditions for the five largest Paraguayan cities from 2016 to 2021.

Usage

py_weather

Format

A data frame with 10945 rows and 28 variables:

rowid

index

country

Country name.

ISO

Country ISO code.

city_name

Character string of the city name.

date

Date, %Y-%m-%d format

tempmax

Maximum temperature at the location.

tempmin

Minimum temperature at the location.

temp

Temperature at the location. Daily values are average values (mean) for the day.

feelslikemax

Maximum feels like temperature at the location.

feelslikemin

Minimum feels like temperature at the location.

feelslike

What the temperature feels like accounting for heat index or wind chill. Daily values are average values (mean) for the day.

dew

Dew point temperature

humidity

Relative humidity in %

precip

The amount of liquid precipitation that fell or is predicted to fall in the period. This includes the liquid-equivalent amount of any frozen precipitation such as snow or ice.

precipprob

The likelihood of measurable precipitation ranging from 0% to 100%

precipcover

The proportion of hours where there was non-zero precipitation. See here

windspeed

The sustained wind speed measured as the average wind speed that occurs during the preceding one to two minutes. Daily values are the maximum hourly value for the day.

winddir

Direction from which the wind is blowing.

sealevelpressure

The sea level atmospheric or barometric pressure in millibars (or hectopascals).

cloudcover

How much of the sky is covered in cloud ranging from 0-100%.

visibility

Distance at which distant objects are visible.

solarradiation

The solar radiation power at the instantaneous moment of the observation (or forecast prediction).

solarenergy

indicates the total energy from the sun that builds up over an hour or day.

uvindex

A value between 0 and 10 indicating the level of ultra violet (UV) exposure for that hour or day. 10 represents high level of exposure, and 0 represents no exposure.

sunrise

The formatted time of the sunrise (For example “2022-05-23T05:50:40”).

sunset

The formatted time of the sunset (For example “2022-05-23T20:22:29”).

moonphase

Represents the fractional portion through the current moon lunation cycle ranging from 0 (the new moon) to 0.5 (the full moon) and back to 1 (the next new moon).

conditions

Textual representation of the weather conditions.

Source

Visual Crossing. https://www.visualcrossing.com/