Weekly crimes of large Paraguayan cities

The py_weekly_crime data frame has 9390 rows and 7 columns. The observations are weekly counts for six different crime types such as theft, aggravated theft, robbery, aggravated robbery, domestic violence, and land transit exposure (accidents). The data correspond to the five largest Paraguayan cities at the moment of data collection: Asuncion(capital), Ciudad del Este, Luque, San Lorenzo, and Capiata.

The definitions of these crimes are according to the Paraguayan Penal Code. The time series cover the years 2016 to 2021.

This data set was used in a published paper analyzing street crime during the first total lockdown in the COVID-19 pandemic era in Paraguay.

Usage

py_weekly_crime

Format

A data frame with 9390 rows and 7 variables:

rowid

index

country

Country name as a character string

ISO

Country ISO code

city

Factor variable of the city name

category

Crime category, factor

week

Reference period. Weekly data, starting on Monday

crimes

Count value corresponding to the respective time period and category

Source

Public Ministry of Paraguay (MP). https://www.ministeriopublico.gov.py/

References

Schneider, A. (2024). Street crime during COVID-19: a retrospective comparison of two Latin American capitals, Revista Paraguay desde las Ciencias Sociales, 14, 73-91, https://publicaciones.sociales.uba.ar/index.php/revistaparaguay/article/view/9800.